Refactor base commands
This pulls what was in the councillor command into the
PerCouncilCommandBase class. This is so that it can be reused.

Along the way I've:
- dropped reporting on skipped/missing/etc councils with the progress
bars - these can come back, but weren't working as intended.
- Made the command runner explicitly call `execute` rather than it
appear in the __init__ method.
- Added a 'pretty' flag to manage what output is shown.
