+++ lgsf/commands/base.py
@@ -4,23 +4,22 @@ import os
 import json
 
 from rich.console import Console
+from rich.progress import Progress, BarColumn, TimeElapsedColumn
 from rich.table import Table
 
 from lgsf.conf import settings
-from lgsf.path_utils import _abs_path, load_scraper
+from lgsf.path_utils import _abs_path, load_scraper, load_council_info
 
 
 class CommandBase(metaclass=abc.ABCMeta):
     command_name = None
 
-    def __init__(self, argv, stdout):
+    def __init__(self, argv, stdout, pretty=False):
         self.argv = argv
         self.create_parser()
         self.stdout = stdout
         self.console = Console(file=self.stdout)
-
-        # After all local vars are set up
-        self.execute()
+        self.pretty = pretty
 
     def create_parser(self):
         self.parser = argparse.ArgumentParser()
@@ -81,8 +80,27 @@ class PerCouncilCommandBase(CommandBase):
             action="store_true",
             help="Only run scrapers not run recently",
         )
+        self.parser.add_argument(
+            "--check-only",
+            action="store_true",
+            help="Just check for updated pages, don't scrape anything",
+        )
+        self.parser.add_argument(
+            "--list-missing",
+            action="store_true",
+            help="Print missing councils",
+        )
+        self.parser.add_argument(
+            "--list-disabled",
+            action="store_true",
+            help="Print disabled councils",
+        )
+
         self.add_default_arguments(self.parser)
-        self.add_arguments(self.parser)
+
+        if hasattr(self, "add_arguments"):
+            self.add_arguments(self.parser)
+
         args = self.parser.parse_args(self.argv[1:])
         if args.list_missing or args.list_disabled:
             return args
@@ -167,6 +185,7 @@ class PerCouncilCommandBase(CommandBase):
             )
         )
 
+    @property
     def councils_to_run(self):
         councils = []
         if self.options["all_councils"] or self.options["tags"]:
@@ -178,6 +197,55 @@ class PerCouncilCommandBase(CommandBase):
                 councils.append(council)
         return councils
 
+    def run_councils(self):
+        for council in self.councils_to_run:
+            self.run_council(council)
+
+    def run_councils_with_progress(self):
+        to_run = self.councils_to_run
+        with Progress(
+            "[progress.description]{task.description}",
+            BarColumn(),
+            "[progress.percentage]{task.percentage:>3.0f}%",
+            TimeElapsedColumn(),
+            console=self.console,
+        ) as progress:
+            total = progress.add_task(description=f"Total", total=len(to_run))
+            while not progress.finished:
+                for council in to_run:
+                    self.run_council(council)
+                    progress.update(total, advance=1)
+
+    def _run_single(self, scraper):
+        try:
+            scraper.run()
+        except KeyboardInterrupt:
+            raise
+        except:
+            if self.options.get("verbose"):
+                raise
+
+    def run_council(self, council):
+        self.options["council"] = council
+        self.options["council_info"] = load_council_info(council)
+        scraper_cls = load_scraper(council, self.command_name)
+        if not scraper_cls:
+            return
+        with scraper_cls(self.options, self.console) as scraper:
+            should_run = True
+            if scraper.disabled:
+                should_run = False
+            if should_run and self.options["refresh"]:
+                if scraper.run_since():
+                    should_run = False
+            if should_run and self.options["tags"]:
+                required_tags = set(self.options["tags"].split(","))
+                scraper_tags = set(scraper.get_tags)
+                if not required_tags.issubset(scraper_tags):
+                    should_run = False
+            if should_run:
+                self._run_single(scraper)
+
     def normalise_codes(self):
         new_codes = []
         if self.options.get("council"):
@@ -187,3 +255,19 @@ class PerCouncilCommandBase(CommandBase):
                 new_codes.append(_abs_path(settings.SCRAPER_DIR_NAME, code)[1])
         self.options["council"] = ",".join(new_codes)
         return self.options
+
+    def handle(self, options):
+        self.options = options
+
+        if options["list_missing"]:
+            self.output_missing()
+
+        if options["list_disabled"]:
+            self.output_disabled()
+
+        self.output_status()
+        self.normalise_codes()
+        if self.pretty:
+            self.run_councils_with_progress()
+        else:
+            self.run_councils()
